api/pkg/github: separate code for importing pull requests
Created with Sturdy
