+++ api/pkg/github/service/module/oss.go
@@ -1,0 +1,13 @@
+//go:build !enterprise && !cloud
+// +build !enterprise,!cloud
+
+package module
+
+import (
+	"getsturdy.com/api/pkg/di"
+	"getsturdy.com/api/pkg/github/service"
+)
+
+func Module(c *di.Container) {
+	c.Register(service.New, new(service.Service))
+}
